+++ lib/after_migrate/collector.rb
@@ -13,24 +13,15 @@ module AfterMigrate
       return unless sql.match?(/\A\s*(CREATE|ALTER|DROP|INSERT|UPDATE|DELETE|RENAME\s+TABLE|TRUNCATE)/i)
 
       table_names = parse_tables(sql)
-      schema ||= fetch_schema
-      # AfterMigrate.log("[#{schema}] Detected change from '#{sql}' to tables: #{table_names}") if table_names.present?
-      collect_tables(schema:, table_names:)
+      schema = fetch_schema
+      AfterMigrate.merge_tables(schema, table_names)
     end
 
     private
 
-    def collect_tables(schema:, table_names:)
-      return if table_names.blank?
-
-      AfterMigrate::Current.affected_tables ||= Hash.new { |h, k| h[k] = Concurrent::Set.new }
-      AfterMigrate::Current.affected_tables[schema].merge(table_names)
-    end
-
     def fetch_schema
       connection = ActiveRecord::Base.connection
-      adapter = connection.adapter_name
-      case adapter
+      case connection.adapter_name
       when 'PostgreSQL'
         quoted = connection.schema_search_path.split(',').first
         quoted&.delete('"')
@@ -39,8 +30,7 @@ module AfterMigrate
 
     def parse_tables(sql)
       connection = ActiveRecord::Base.connection
-      adapter = connection.adapter_name
-      case adapter
+      case connection.adapter_name
       when 'PostgreSQL'
         AfterMigrate::Postgresql.parse_tables(sql)
       when 'SQLite'
@@ -48,7 +38,7 @@ module AfterMigrate
       when 'Mysql2', 'Trilogy'
         AfterMigrate::Mysql.parse_tables(sql)
       else
-        AfterMigrate.log("No maintenance implemented for #{adapter}")
+        AfterMigrate.log("No maintenance implemented for #{connection.adapter_name}")
       end
     end
   end